chore: Fix clippy.

+++ core/src/database/tx_sender.rs
@@ -145,7 +145,7 @@ impl Database {
             AND txs.seen_block_id IS NULL",
                 common_ctes
             ))
-            .bind(block_id as i32)
+            .bind(block_id)
             .fetch_all(&bg_db.connection)
             .await
             else {
@@ -162,7 +162,7 @@ impl Database {
             AND txs.seen_block_id IS NULL",
                 common_ctes
             ))
-            .bind(block_id as i32)
+            .bind(block_id)
             .fetch_all(&bg_db.connection)
             .await
             else {
